Create mule.mule module (split out from driver.py)
Renamed {driver,main}.py
